requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==41.0.7
cymem==2.0.8
dataclasses-json==0.6.3
distlib==0.3.8
distro==1.9.0
Django==5.0
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
errant==3.0.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
googletrans==4.0.0rc1
greenlet==3.0.3
gTTS==2.5.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
htbuilder==0.6.2
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.11.0
inexactsearch==1.0.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langcodes==3.3.0
langsmith==0.0.77
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
openai==0.28.1
packaging==23.2
pandas==2.1.4
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
platformdirs==4.1.0
preshed==3.0.9
protobuf==4.25.1
pyarrow==14.0.2
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pypdfium2==4.25.0
pyspellchecker==0.7.3
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.16.2
safetensors==0.4.1
silpa_common==0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundex==1.1.3
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
# spellchecker==0.5.6
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==2.4.8
st-annotated-text==4.0.1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.4
yarl==1.9.4
zipp==3.17.0
